refactor(users): Move and adapt UserService to domain
