+++ apps/backend/users-service/src/users/application/index.ts
@@ -1,0 +1,1 @@
+export * from './users.service';
